fixed employee removing problem

+++ backend/routes/userRoutes.js
@@ -1,9 +1,11 @@
 const express = require("express")
 const bcrypt = require('bcrypt');
 const User = require("../models/userModel");
+const Organization = require("../models/organizationModel")
 const { ObjectId } = require("mongodb");
 const { generateRandomPassword } = require("../utils/generators");
-const { sendMail } = require("../utils/mail")
+const { sendMail } = require("../utils/mail");
+
 const router = express.Router()
 
 const saltRounds = 10;
@@ -34,28 +36,35 @@ router.post("/createUser", async (req, res) => {
 
         const newUser = await User.create(userData);
 
-        sendMail(
-            "Stockwise Admin",
-            userData["email"],
-            "Welcome to Stockwise",
-            "",
-            `
-            <h3>We welcome you to the Stockwise</h3>
-            <p>
+        try {
+
+            sendMail(
+                "Stockwise Admin",
+                userData["email"],
+                "Welcome to Stockwise",
+                "",
+                `
+                <h3>We welcome you to the Stockwise</h3>
+                <p>
                 You have been added to the Organization: <br>
                 as a ${userData["role"]} <br>
                 As a next step <br>
                 We recommend you to login to the account change your password
-            </p>
+                </p>
                 <strong>Here are your credentials<strong>
                 email: ${userData["email"]}
                 password: ${randomPassword}
-
-            <h4>
+                
+                <h4>
                 We hope you have a long and lovely relationship with Stockwise
-            </h4>
-            `
-        )
+                </h4>
+                `
+            )
+        }
+        catch (error) {
+            console.log("Could not send mail", error);
+        }
+
         res.json(newUser);
     } catch (error) {
         res.status(500).json({ message: "Internal Server Error. Could not create user." });
@@ -91,14 +100,25 @@ router.delete("/deleteUser", async (req, res) => {
 
     try {
         const existingUser = await User.findOne({ _id: new Object(_id) });
+        const org = await Organization.findOne({ employees: new ObjectId(existingUser._id)});
+        
         if (!existingUser) {
             return res.status(404).json({ message: "User not found" });
         }
+
+        if (!org) {
+            return res.status(404).json({ error: `Organization with orgId: ${existingUser._id} does not exist` });
+        }
+
+        // remove employee from the org's employees array
+        org.employees = org.employees.filter(id => id.toString() !== existingUser._id.toString());
+        await org.save();
 
         await User.findOneAndDelete({ _id: new Object(_id) });
 
         res.json({ message: "User deleted successfully" });
     } catch (error) {
+        console.log("error: ", error)
         res.status(500).json({ message: "Internal Server Error. Could not delete user." });
     }
 });
